feat: use React Router to support pages
