+++ frontend/package.json
@@ -10,6 +10,7 @@
     "react-chord-diagram": "^1.5.0",
     "react-dom": "^16.10.2",
     "react-dropzone-uploader": "^2.10.2",
+    "react-router-dom": "^5.1.2",
     "react-scripts": "3.2.0",
     "react-toastify": "^5.4.0",
     "socket.io-client": "^2.3.0"
